requirements.txt
backports.functools-lru-cache==1.5
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.2
Django==1.11.20
django-cors-headers==2.2.0
djangorestframework==3.9.1
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.8
jsonpickle==1.1
mysql-connector-python==8.0.13
mysqlclient==1.3.13
oauth2client==4.1.3
oauthlib==3.0.1
Pillow==5.4.1
protobuf==3.6.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytube==9.4.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
soupsieve==1.7.3
uritemplate==3.0.0
urllib3==1.24.1
